refactor(VThemeProvider): add unit tests, clean-up theme page
